+++ app/packages/partup-client-dropbox-chooser/DropboxChooser.ctrl.js
@@ -94,6 +94,7 @@ if (Meteor.isClient) {
                                 path: '/' + decodeURI(path[2]),
                                 settings: {
                                     "requested_visibility": "public"
+                                    // "expires": "2045-05-12T15:50:38Z" --> This is only allowed for paying dropbox users.
                                 }
                             }),
                         })