+++ app/packages/partup-client-dropdowns/chats/chat-notifications.js
@@ -1,107 +1,144 @@
-Template.DropdownChatNotifications.onCreated(function() {
+Template.DropdownChatNotifications.onCreated(function () {
     var template = this;
-    template.private = new ReactiveVar(true);
+    template.isPrivateChat = new ReactiveVar(true);
+    template.newMessage = new ReactiveVar(undefined);
+    template.latestMessage = new ReactiveVar(0);
+
+    /** variable name gets used by the ClientDropdowns event handler. */
     template.dropdownOpen = new ReactiveVar(false);
-    template.setDropdownState = function(state) {
-        template.dropdownOpen.set(state);
-    };
-    template.subscribe('chats.for_loggedin_user', {networks: true, private: true}, {});
+
+    template.subscribe('chats.for_loggedin_user', { networks: true, private: true }, {});
 });
-Template.DropdownChatNotifications.onRendered(function() {
+Template.DropdownChatNotifications.onRendered(function () {
     var template = this;
-    ClientDropdowns.addOutsideDropdownClickHandler(template, '[data-clickoutside-close]', '[data-toggle-menu=chat-notifications]', function() {ClientDropdowns.partupNavigationSubmenuActive.set(false);});
-    Router.onBeforeAction(function(req, res, next) {
+    ClientDropdowns.addOutsideDropdownClickHandler(template, '[data-clickoutside-close]', '[data-toggle-menu=chat-notifications]', function () { ClientDropdowns.partupNavigationSubmenuActive.set(false); });
+    Router.onBeforeAction(function (req, res, next) {
         template.dropdownOpen.set(false);
         next();
     });
 });
 
-Template.DropdownChatNotifications.onDestroyed(function() {
+Template.DropdownChatNotifications.onDestroyed(function () {
     var template = this;
     ClientDropdowns.removeOutsideDropdownClickHandler(template);
 });
 
 Template.DropdownChatNotifications.events({
     'DOMMouseScroll [data-preventscroll], mousewheel [data-preventscroll]': Partup.client.scroll.preventScrollPropagation,
-    'click [data-toggle-menu]': ClientDropdowns.dropdownClickHandler.bind(null, 'top-level'),
-    'click [data-private]': function(event, template) {
-        event.preventDefault();
-        template.private.set(true);
+    'click [data-toggle-menu]': function (event, template) {
+        template.newMessage.set(false);
+        ClientDropdowns.dropdownClickHandler(event, template);
+        //ClientDropdowns.dropdownClickHandler.bind(null, 'top-level') <-- old implementation.
     },
-    'click [data-network]': function(event, template) {
+    'click [data-private]': function (event, template) {
         event.preventDefault();
-        template.private.set(false);
+        template.isPrivateChat.set(true);
+    },
+    'click [data-network]': function (event, template) {
+        event.preventDefault();
+        template.isPrivateChat.set(false);
     }
 });
 
 Template.DropdownChatNotifications.helpers({
-    menuOpen: function() {
+    dropdownOpen: function () {
         return Template.instance().dropdownOpen.get();
     },
-    reactiveDropdownOpen: function() {
-        return Template.instance().dropdownOpen;
+    isPrivateChat: function () {
+        return Template.instance().isPrivateChat.get();
     },
-    private: function() {
-        return Template.instance().private.get();
-    },
-    data: function() {
+    chatData: function () {
         var template = Template.instance();
         var userId = Meteor.userId();
+
         if (!userId) return;
-        var getPrivate = template.private.get();
         var currentActiveChatId = Session.get('partup-current-active-chat');
 
-        var networkChats = Chats.findForUser(userId, {networks: true}, {fields: {_id: 1, counter: 1, created_at: 1}, sort: {updated_at: -1}})
-            .map(function(chat) {
-                chat.message = ChatMessages.findOne({chat_id: chat._id}, {sort: {created_at: -1}, limit: 1});
+        var privateChats =
+            Chats.findForUser(userId, { private: true }, { fields: { _id: 1, counter: 1, created_at: 1 }, sort: { updated_at: -1 } })
+                .map(function (chat) {
+                    chat.creator =
+                        Meteor.users.findOne({ _id: { $nin: [userId] }, chats: { $in: [chat._id] } });
+                    chat.message =
+                        ChatMessages.findOne({ chat_id: chat._id }, { sort: { created_at: -1 }, limit: 1 });
+                    return chat;
+                });
+        var networkChats =
+            Chats.findForUser(userId, { networks: true }, { fields: { _id: 1, counter: 1, created_at: 1 }, sort: { updated_at: -1 } })
+                .map(function (chat) {
+                    chat.message =
+                        ChatMessages.findOne({ chat_id: chat._id }, { sort: { created_at: -1 }, limit: 1 });
+                    if (chat.message) {
+                        chat.message.creator =
+                            Meteor.users.findOne({ _id: chat.message.creator_id });
+                    }
+                    return chat;
+                });
 
-                if (chat.message) chat.message.creator = Meteor.users.findOne({_id: chat.message.creator_id});
+        var totalChatMessages = function (chatCollection) {
+            return chatCollection
+                .map(function (chat) {
+                    return chat._id === currentActiveChatId ? 0 : chat.unreadCount();
+                }).reduce(function (prev, curr) {
+                    return prev + curr;
+                }, 0);
+        };
+        var chatMessageTime = function (chat) {
+            return chat.message ? Date.parse(chat.message.created_at) : 0;
+        };
+        var latestChat = function () {
+            var getLatest = function (chatCollection) {
+                return _.max(
+                    _.filter(chatCollection, function (chat) {
+                        return (chat.message && chat.message.creator_id !== userId);
+                    }), function (chat) {
+                        return chatMessageTime(chat);
+                    });
+            };
+            var private = getLatest(privateChats);
+            var network = getLatest(networkChats);
 
-                return chat;
-            });
-
-        var privateChats = Chats.findForUser(userId, {private: true}, {fields: {_id: 1, counter: 1, created_at: 1}, sort: {updated_at: -1}})
-            .map(function(chat) {
-                chat.creator = Meteor.users
-                    .findOne({_id: {$nin: [userId]}, chats: {$in: [chat._id]}});
-
-                chat.message = ChatMessages.findOne({chat_id: chat._id}, {sort: {created_at: -1}, limit: 1});
-
-                return chat;
-            });
+            return chatMessageTime(private) > chatMessageTime(network) ? private : network;
+        };
+        var hasNewMessage = function (chat, messageTime, unreadMessageCount) {
+                if (template.newMessage.get() === undefined) {
+                    if (unreadMessageCount) {
+                        template.newMessage.set(true);
+                    }
+                    template.latestMessage.set(messageTime);
+                } else {
+                    if (chat && chat._id === currentActiveChatId) {
+                        template.newMessage.set(false);
+                        template.latestMessage.set(messageTime);
+                    } else {
+                        if (messageTime > template.latestMessage.get()) {
+                            template.newMessage.set(true);
+                            template.latestMessage.set(messageTime);
+                        }
+                    }
+                }
+                return template.newMessage.get();
+        };
 
         return {
-            chats: function() {
-                if (getPrivate) return privateChats;
-                return networkChats;
-            },
-            totalTribeMessages: function() {
-                return networkChats
-                    .map(function(chat) {
-                        if (chat._id === currentActiveChatId) return 0;
-                        return chat.unreadCount();
-                    }).reduce(function(prev, curr) {
-                        return prev + curr;
-                    }, 0);
-            },
-            totalPersonalMessages: function() {
-                return privateChats
-                    .map(function(chat) {
-                        if (chat._id === currentActiveChatId) return 0;
-                        return chat.unreadCount();
-                    }).reduce(function(prev, curr) {
-                        return prev + curr;
-                    }, 0);
+            chats: template.isPrivateChat.get() ? privateChats : networkChats,
+            totalPrivateMessages: totalChatMessages(privateChats),
+            totalNetworkMessages: totalChatMessages(networkChats),
+            hasNewMessages: function () {
+                var chat = latestChat();
+                var messageTime = chatMessageTime(chat);
+                var unreadMessageCount = this.totalPrivateMessages || this.totalNetworkMessages;
+                return hasNewMessage(chat, messageTime, unreadMessageCount);
             }
-        };
+        }
     },
-    notificationClickHandler: function() {
+    notificationClickHandler: function () {
         var template = Template.instance();
-        return function() {
-            template.setDropdownState(false);
+        return function () {
+            template.dropdownOpen.set(false);
         };
     },
-    appStoreLink: function() {
+    appStoreLink: function () {
         return Partup.client.browser.getAppStoreLink();
     }
 });