+++ app/packages/partup-lib/services/validators.js
@@ -5,7 +5,7 @@
  */
 Partup.services.validators = {
     // tagsSeparatedByComma: /^\s*(\w|-|&|\.)*(\s*,?\s*(\w|-|&|\.)*)*\s*$/, // old
-    tagsSeparatedByComma: /^\s*(\w|-|&|\.|[^\x00-\x7F\w])*(\s*,?\s*(\w|-|&|\.|[^\x00-\x7F\w])*)*\s*$/, // new with unicode support
+    tagsSeparatedByComma: /^\s*(\w|-|&|\.|['^\x00-\x7F\w])*(\s*,?\s*(\w|-|&|\.|['^\x00-\x7F\w])*)*\s*$/, // new with unicode support
 
     // minimum 8 characters, at least 1 number, at least 1 capital letter
     password: /(?=^.{8,}$)(?=.*\d)(?![.\n])(?=.*[A-Z])(?=.*[a-z]).*$/,